requirements.txt
flask
boto3
python-dotenv
Bootstrap-Flask==2.4.2

